minio-datalake/kafka-consumer/requirements.txt
kafka-python
Pillow
boto3